requirements.txt
aiogram==3.3.0
asyncpg==0.29.0
aiosqlite==0.19.0
sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1
httpx==0.27.0
yt-dlp==2024.4.9
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9